# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lander Dispatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: List transactions
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /transactions/{transactionId}:
    get:
      tags: [Transactions]
      operationId: getTransaction
      summary: Get a transaction
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    patch:
      tags: [Transactions]
      operationId: updateTransaction
      summary: Update a transaction
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    delete:
      tags: [Transactions]
      operationId: deleteTransaction
      summary: Delete a transaction
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
components:
...
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # Dashboard
...
    Transaction:
      type: object
      required: [id, transactionId, type, category, amount, status, createdAt]
      properties:
        id:
          type: string
        transactionId:
          type: string
        date:
          type: [string, 'null']
          format: date
        dueDate:
          type: [string, 'null']
          format: date
        type:
          type: string
        category:
          type: string
        description:
          type: [string, 'null']
        carrierId:
          type: [string, 'null']
        carrierName:
          type: [string, 'null']
        invoiceId:
          type: [string, 'null']
        invoiceNumber:
          type: [string, 'null']
        paymentMethod:
          type: [string, 'null']
        amount:
          type: number
        status:
          type: string
        referenceNumber:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    TransactionInput:
      type: object
      required: [type, category, amount]
      properties:
        date:
          type: string
        dueDate:
          type: string
        type:
          type: string
        category:
          type: string
        description:
          type: string
        carrierId:
          type: string
        invoiceId:
          type: string
        paymentMethod:
          type: string
        amount:
          type: number
        status:
          type: string
        referenceNumber:
          type: string
        notes:
          type: string
    TransactionUpdate:
      type: object
      properties:
        date:
          type: string
        dueDate:
          type: string
        type:
          type: string
        category:
          type: string
        description:
          type: string
        carrierId:
          type: string
        invoiceId:
          type: string
        paymentMethod:
          type: string
        amount:
          type: number
        status:
          type: string
        referenceNumber:
          type: string
        notes:
          type: string
    TransactionListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...